# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetricFilter": {
        "additionalProperties": false,
        "description": "One comparison against a field the metric declares as filterable.",
        "properties": {
          "field": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Field",
            "type": "string"
          },
          "operator": {
            "default": "eq",
            "enum": [
              "eq",
              "ne",
              "in",
              "not_in",
              "gt",
              "gte",
              "lt",
              "lte"
            ],
            "title": "Operator",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "type": "array"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "field",
          "value"
        ],
        "title": "MetricFilter",
        "type": "object"
      },
...
      "MetricParameters": {
        "additionalProperties": false,
        "description": "Everything a reader may change about one factual section of a report.",
        "properties": {
          "dimensions": {
            "items": {
              "type": "string"
            },
            "maxItems": 3,
            "title": "Dimensions",
            "type": "array"
          },
          "filters": {
            "items": {
              "$ref": "#/components/schemas/MetricFilter"
            },
            "maxItems": 8,
            "title": "Filters",
            "type": "array"
          },
          "grain": {
            "default": "month",
            "enum": [
              "day",
              "week",
              "month",
              "quarter",
              "year"
            ],
            "title": "Grain",
            "type": "string"
          },
          "metric": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Metric",
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/ReportPeriod"
          }
        },
        "required": [
          "metric",
          "period"
        ],
        "title": "MetricParameters",
        "type": "object"
      },
...
      "PublishReportRequest": {
        "additionalProperties": false,
        "description": "What a caller may choose when publishing. Never what a report says.\n\nRejecting unknown fields rather than ignoring them is the point: a request\ncarrying a figure would otherwise look accepted, and a caller could believe\nit had put a number into a report. Every value in a published document comes\nfrom a query the run executed.",
        "properties": {
          "formats": {
            "items": {
              "enum": [
                "pdf",
                "docx"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Formats",
            "type": "array"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricParameters"
            },
            "maxItems": 8,
            "title": "Metrics",
            "type": "array"
          },
          "narrative": {
            "anyOf": [
              {
                "enum": [
                  "current",
                  "pinned_to_original_period",
                  "excluded_from_refreshed_report"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative"
          },
          "period": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "template": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Template",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "template",
          "formats"
        ],
        "title": "PublishReportRequest",
        "type": "object"
      },
      "PublishReportResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/PublishedDocumentResponse"
            },
            "title": "Documents",
            "type": "array"
          },
          "narrative": {
            "default": "current",
            "enum": [
              "current",
              "pinned_to_original_period",
              "excluded_from_refreshed_report"
            ],
            "title": "Narrative",
            "type": "string"
          },
          "rerun_query_ids": {
            "items": {
              "type": "string"
            },
            "title": "Rerun Query Ids",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "template": {
            "title": "Template",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "template",
          "documents"
        ],
        "title": "PublishReportResponse",
        "type": "object"
      },
      "PublishedDocumentResponse": {
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "artifact_id",
          "name",
          "media_type",
          "size"
        ],
        "title": "PublishedDocumentResponse",
        "type": "object"
      },
...
      "ReportPeriod": {
        "additionalProperties": false,
        "description": "A half-open date range, ``start`` inclusive and ``end`` exclusive.\n\nHalf-open because a closed range over timestamps either drops the last day's\nevening or double-counts a boundary, depending on how it is written, and\nboth mistakes are invisible in a published total.",
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "ReportPeriod",
        "type": "object"
      },
...
      "ScheduleConfigPayload": {
        "additionalProperties": false,
        "properties": {
          "day_of_month": {
            "anyOf": [
              {
                "maximum": 28.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Month"
          },
          "day_of_week": {
            "anyOf": [
              {
                "maximum": 6.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Week"
          },
          "hour": {
            "default": 6,
            "maximum": 23.0,
            "minimum": 0.0,
            "title": "Hour",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "quarterly"
            ],
            "title": "Kind",
            "type": "string"
          },
          "minute": {
            "default": 0,
            "maximum": 59.0,
            "minimum": 0.0,
            "title": "Minute",
            "type": "integer"
          },
          "month_of_quarter": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month Of Quarter"
          }
        },
        "required": [
          "kind"
        ],
        "title": "ScheduleConfigPayload",
        "type": "object"
      },
      "ScheduledReportCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "delivery_channel": {
            "anyOf": [
              {
                "enum": [
                  "link",
                  "webhook",
                  "email"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Channel"
          },
          "delivery_destination": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Destination"
          },
          "formats": {
            "items": {
              "enum": [
                "pdf",
                "docx"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "title": "Formats",
            "type": "array"
          },
          "saved_report_id": {
            "format": "uuid",
            "title": "Saved Report Id",
            "type": "string"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleConfigPayload"
          },
          "timezone": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          },
          "workspace_id": {
            "default": "default",
            "maxLength": 128,
            "minLength": 1,
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "saved_report_id",
          "schedule",
          "timezone"
        ],
        "title": "ScheduledReportCreateRequest",
        "type": "object"
      },
      "ScheduledReportListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScheduledReportResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ScheduledReportListResponse",
        "type": "object"
      },
      "ScheduledReportResponse": {
        "properties": {
          "consecutive_failures": {
            "title": "Consecutive Failures",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivery_channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Channel"
          },
          "delivery_destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Destination"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "formats": {
            "items": {
              "type": "string"
            },
            "title": "Formats",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Result"
          },
          "last_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "next_run_at": {
            "format": "date-time",
            "title": "Next Run At",
            "type": "string"
          },
          "saved_report_id": {
            "title": "Saved Report Id",
            "type": "string"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleConfigPayload"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "saved_report_id",
          "workspace_id",
          "schedule",
          "timezone",
          "formats",
          "delivery_channel",
          "delivery_destination",
          "enabled",
          "next_run_at",
          "last_run_at",
          "last_result",
          "consecutive_failures",
          "created_at",
          "updated_at"
        ],
        "title": "ScheduledReportResponse",
        "type": "object"
      },
      "ScheduledReportUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "delivery_channel": {
            "anyOf": [
              {
                "enum": [
                  "link",
                  "webhook",
                  "email"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Channel"
          },
          "delivery_destination": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Destination"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "formats": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "pdf",
                    "docx"
                  ],
                  "type": "string"
                },
                "maxItems": 2,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formats"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleConfigPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "ScheduledReportUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Autonomous Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analytics/runs/{run_id}/reports": {
      "post": {
        "description": "Assemble a completed run into documents. This never calls the model.",
        "operationId": "publish_report_api_v1_analytics_runs__run_id__reports_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Report",
        "tags": [
          "analytics-workbench"
        ]
      }
    },
...
    "/api/v1/reports/scheduled": {
      "get": {
        "operationId": "list_scheduled_reports_api_v1_reports_scheduled_get",
        "parameters": [
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enabled"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledReportListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scheduled Reports",
        "tags": [
          "scheduled-reports"
        ]
      },
      "post": {
        "operationId": "create_scheduled_report_api_v1_reports_scheduled_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledReportCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scheduled Report",
        "tags": [
          "scheduled-reports"
        ]
      }
    },
    "/api/v1/reports/scheduled/{scheduled_report_id}": {
      "get": {
        "operationId": "get_scheduled_report_api_v1_reports_scheduled__scheduled_report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scheduled_report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scheduled Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scheduled Report",
        "tags": [
          "scheduled-reports"
        ]
      },
      "patch": {
        "operationId": "update_scheduled_report_api_v1_reports_scheduled__scheduled_report_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "scheduled_report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scheduled Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledReportUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Scheduled Report",
        "tags": [
          "scheduled-reports"
        ]
      }
    },
...
  }
}